Change how last_updated_commit_id is implemented
The previous implementation was slow and clunky, this one is way faster (because it avoid running through the complete history graph) and *shoulb* be less clunky.
It has been tested successfully on various huge repositories (such as GHC) with various weird cases, but I'm not sure it is 100% perfect.
Either way, this should work better than the current one.
Fixes #33